Refactor(authentication) : 패키지 정리
Related to : #41
