+++ src/main/java/org/deco/gachicoding/config/SecurityConfig.java
@@ -1,5 +1,11 @@
 package org.deco.gachicoding.config;
 
+import org.deco.gachicoding.config.security.RestLoginAuthenticationEntryPoint;
+import org.deco.gachicoding.config.security.RestAuthenticationProvider;
+import org.deco.gachicoding.config.security.RestLoginProcessingFilter;
+import org.deco.gachicoding.config.security.handler.RestAccessDeniedHandler;
+import org.deco.gachicoding.config.security.handler.RestAuthenticationFailureHandler;
+import org.deco.gachicoding.config.security.handler.RestAuthenticationSuccessHandler;
 import org.deco.gachicoding.user.application.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
