+++ src/main/java/org/deco/gachicoding/config/security/RestAuthenticationProvider.java
@@ -1,8 +1,7 @@
-package org.deco.gachicoding.config;
+package org.deco.gachicoding.config.security;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.deco.gachicoding.user.application.UserAuthenticationDto;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.Authentication;
@@ -20,12 +19,17 @@ public class RestAuthenticationProvider implements AuthenticationProvider {
 
     @Override
     public Authentication authenticate(Authentication authentication) throws AuthenticationException {
+
         String loginId = (String) authentication.getPrincipal();
         String password = (String) authentication.getCredentials();
+
         log.info("loginId - {}", loginId);
         log.info("password - {}", password);
+
         UserDetails userDto = userDetailsService.loadUserByUsername(loginId);
+
         log.info("{}", userDto.toString());
+
         if (!passwordEncoder.matches(password, userDto.getPassword())) {
             throw new BadCredentialsException("Invalid password");
         }
